register tpmr as a bridge port, not an input handler, on member ifaces.
this is a step toward making all types of bridges coordinate their
use of port interfaces, and is a step toward deprecating the interface
input handler lists. it also moves tpmr away from the trunk ioctls
it's currently (ab)using.

this has been in snaps as part of a larger diff for over a week.