setitimer(2): realitexpire(): call getnanouptime(9) once
timespecadd(3) is fast.  There is no need to call getnanouptime(9)
repeatedly when searching for the next expiration point.  Given that
it_interval is at least 1/hz, we expect to run through the loop maybe
hz times at most.  Even at HZ=10000 that's pretty brief.

While we're here, pull *all* of the other logic out of the loop.
The only thing we need to do in the loop is timespecadd(3).