bpf(4): remove ticks
Change bd_rtout to a uint64_t of nanoseconds.  Update the code in
bpfioctl() and bpfread() accordingly.

Add a local copy of nsecuptime() to make the diff smaller.  This will
need to move to kern_tc.c if/when we have another user elsewhere in
the kernel.

Prompted by mpi@.  With input from dlg@.

ok dlg@ mpi@ visa@