Let a process open a video(4) device multiple times
The previous behavior was conservative but a bit too restrictive.
The V4L API lets several processes open a single video device, with
exclusive access to certain methods when a process starts streaming.
This is not trivial to implement.

This small change fixes webcam usage with Firefox and BigBlueButton.
Diff simplified + ok mglocker@, thanks!