more consistently use pfsync_free_deferral to free the mbuf.
pfsync_free_deferral doesnt need to check pd_m for NULL before
calling m_freem because m_freem does that anyway.

if pf_setup_pdesc in pfsync_undefer_notify failed, the mbuf was
freed but the pd_m pointer was not cleared, which would have led
to a double free when pfsync_free_deferral tried to do the same
thing for it.

if pfsync_undefer is supposed to drop the mbuf, let pfsync_free_deferral
do it for us.

ok jmatthew@